+++ integrations/jira/main.py
@@ -10,6 +10,7 @@ from port_ocean.core.ocean_types import ASYNC_GENERATOR_RESYNC_TYPE
 class ObjectKind(StrEnum):
     PROJECT = "project"
     ISSUE = "issue"
+    USER = "user"
 
 
 async def setup_application() -> None:
@@ -59,6 +60,19 @@ async def on_resync_issues(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
         yield issues
 
 
+@ocean.on_resync(ObjectKind.USER)
+async def on_resync_users(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
+    client = JiraClient(
+        ocean.integration_config["jira_host"],
+        ocean.integration_config["atlassian_user_email"],
+        ocean.integration_config["atlassian_user_token"],
+    )
+
+    async for users in client.get_paginated_users():
+        logger.info(f"Received users batch with {len(users)} users")
+        yield users
+
+
 @ocean.router.post("/webhook")
 async def handle_webhook_request(data: dict[str, Any]) -> dict[str, Any]:
     client = JiraClient(
@@ -66,16 +80,51 @@ async def handle_webhook_request(data: dict[str, Any]) -> dict[str, Any]:
         ocean.integration_config["atlassian_user_email"],
         ocean.integration_config["atlassian_user_token"],
     )
-    logger.info(f'Received webhook event of type: {data.get("webhookEvent")}')
-    if "project" in data:
-        logger.info(f'Received webhook event for project: {data["project"]["key"]}')
-        project = await client.get_single_project(data["project"]["key"])
-        await ocean.register_raw(ObjectKind.PROJECT, [project])
-    elif "issue" in data:
-        logger.info(f'Received webhook event for issue: {data["issue"]["key"]}')
-        issue = await client.get_single_issue(data["issue"]["key"])
-        await ocean.register_raw(ObjectKind.ISSUE, [issue])
-    logger.info("Webhook event processed")
+
+    webhook_event = data.get("webhookEvent")
+    if not webhook_event:
+        logger.error("Missing webhook event")
+        return {"ok": False, "error": "Missing webhook event"}
+
+    logger.info(f"Processing webhook event: {webhook_event}")
+
+    match webhook_event:
+        case event if event.startswith("user_"):
+            account_id = data["user"]["accountId"]
+            logger.debug(f"Fetching user with accountId: {account_id}")
+            item = await client.get_single_user(account_id)
+            kind = ObjectKind.USER
+        case event if event.startswith("project_"):
+            project_key = data["project"]["key"]
+            logger.debug(f"Fetching project with key: {project_key}")
+            item = await client.get_single_project(project_key)
+            kind = ObjectKind.PROJECT
+        case event if event.startswith("jira:issue_"):
+            issue_key = data["issue"]["key"]
+            logger.debug(f"Fetching issue with key: {issue_key}")
+            item = await client.get_single_issue(issue_key)
+            kind = ObjectKind.ISSUE
+        case _:
+            logger.error(f"Unknown webhook event type: {webhook_event}")
+            return {
+                "ok": False,
+                "error": f"Unknown webhook event type: {webhook_event}",
+            }
+
+    if not item:
+        logger.error("Failed to retrieve item")
+        return {"ok": False, "error": "Failed to retrieve item"}
+
+    logger.debug(f"Retrieved {kind} item: {item}")
+
+    if "deleted" in webhook_event:
+        logger.info(f"Unregistering {kind} item")
+        await ocean.unregister_raw(kind, [item])
+    else:
+        logger.info(f"Registering {kind} item")
+        await ocean.register_raw(kind, [item])
+
+    logger.info(f"Webhook event '{webhook_event}' processed successfully")
     return {"ok": True}